feat(webgl): Support WebGPU texture formats
