+++ modules/webgl/src/classes/clear.ts
@@ -1,6 +1,7 @@
 import {Device, assert} from '@luma.gl/api';
 import {assertWebGL2Context} from '../context/context/webgl-checks';
 import {withParameters} from '../context/state-tracker/with-parameters';
+import WebGLDevice from '../adapter/webgl-device';
 import Framebuffer from './framebuffer';
 
 // Should collapse during minification
@@ -18,11 +19,10 @@ const ERR_ARGUMENTS = 'clear: bad arguments';
 
 /** Optionally clears depth, color and stencil buffers */
 export function clear(
-  device: Device | WebGLRenderingContext,
+  gl: Device | WebGLRenderingContext,
   options?: {framebuffer?: Framebuffer; color?: any; depth?: any; stencil?: any}
 ): void {
-  // @ts-expect-error Extract context
-  const gl = device.gl || device;
+  const device = WebGLDevice.attach(gl);
   const {framebuffer = null, color = null, depth = null, stencil = null} = options || {};
   const parameters: any = {};
 
@@ -56,49 +56,47 @@ export function clear(
   assert(clearFlags !== 0, ERR_ARGUMENTS);
 
   // Temporarily set any clear "colors" and call clear
-  withParameters(gl, parameters, () => {
-    gl.clear(clearFlags);
+  withParameters(device.gl, parameters, () => {
+    device.gl.clear(clearFlags);
   });
 }
 
 /** WebGL2 - clear a specific drawing buffer */
 export function clearBuffer(
-  device: Device | WebGLRenderingContext,
+  gl: Device | WebGLRenderingContext,
   options?: {framebuffer?: Framebuffer; buffer?: any; drawBuffer?: any; value?: any}
 ) {
-  // @ts-expect-error Extract context
-  const gl = device.gl || device;
-  assertWebGL2Context(gl);
+  const device = WebGLDevice.attach(gl);
 
   const {framebuffer = null, buffer = GL_COLOR, drawBuffer = 0, value = [0, 0, 0, 0]} = options || {};
-  withParameters(gl, {framebuffer}, () => {
+  withParameters(device.gl2, {framebuffer}, () => {
     // Method selection per OpenGL ES 3 docs
     switch (buffer) {
       case GL_COLOR:
         switch (value.constructor) {
           case Int32Array:
-            gl.clearBufferiv(buffer, drawBuffer, value);
+            device.gl2.clearBufferiv(buffer, drawBuffer, value);
             break;
           case Uint32Array:
-            gl.clearBufferuiv(buffer, drawBuffer, value);
+            device.gl2.clearBufferuiv(buffer, drawBuffer, value);
             break;
           case Float32Array:
           default:
-            gl.clearBufferfv(buffer, drawBuffer, value);
+            device.gl2.clearBufferfv(buffer, drawBuffer, value);
         }
         break;
 
       case GL_DEPTH:
-        gl.clearBufferfv(GL_DEPTH, 0, [value]);
+        device.gl2.clearBufferfv(GL_DEPTH, 0, [value]);
         break;
 
       case GL_STENCIL:
-        gl.clearBufferiv(GL_STENCIL, 0, [value]);
+        device.gl2.clearBufferiv(GL_STENCIL, 0, [value]);
         break;
 
       case GL_DEPTH_STENCIL:
         const [depth, stencil] = value;
-        gl.clearBufferfi(GL_DEPTH_STENCIL, 0, depth, stencil);
+        device.gl2.clearBufferfi(GL_DEPTH_STENCIL, 0, depth, stencil);
         break;
 
       default:
