+++ modules/webgl/src/classes/texture-2d.ts
@@ -1,17 +1,21 @@
 import {Device, loadImage, TextureData} from '@luma.gl/api';
 import GL from '@luma.gl/constants';
-import Texture, {TextureProps, TextureSupportOptions} from './texture';
+import Texture, {TextureProps} from './texture';
 
 export type Texture2DProps = TextureProps & {
   format?: number;
 };
 
+/**
+ * Standard 2D Texture
+ * @deprecated Use device.createTexture()
+ */
 export default class Texture2D extends Texture {
   // get [Symbol.toStringTag](): string {
   //   return 'Texture2D';
   // }
 
-  static isSupported(device: Device | WebGLRenderingContext, opts?: TextureSupportOptions): boolean {
+  static isSupported(device: Device | WebGLRenderingContext, opts?): boolean { // : TextureSupportOptions
     return Texture.isSupported(device, opts);
   }
