+++ modules/webgl/src/types/webgl.ts
@@ -9,11 +9,14 @@ export type WebGLSamplerParameters = {
   [GL.TEXTURE_WRAP_T]?: GL.CLAMP_TO_EDGE | GL.REPEAT | GL.MIRRORED_REPEAT;
   [GL.TEXTURE_WRAP_R]?: GL.CLAMP_TO_EDGE | GL.REPEAT | GL.MIRRORED_REPEAT;
   [GL.TEXTURE_MAG_FILTER]?: GL.NEAREST | GL.LINEAR;
-  [GL.TEXTURE_MIN_FILTER]?: GL.NEAREST_MIPMAP_NEAREST | GL.NEAREST_MIPMAP_LINEAR | GL.LINEAR_MIPMAP_NEAREST | GL.LINEAR_MIPMAP_LINEAR;
+  [GL.TEXTURE_MIN_FILTER]?:
+    GL.NEAREST | GL.LINEAR |
+    GL.NEAREST_MIPMAP_NEAREST | GL.NEAREST_MIPMAP_LINEAR |
+    GL.LINEAR_MIPMAP_NEAREST | GL.LINEAR_MIPMAP_LINEAR;
   [GL.TEXTURE_MIN_LOD]?: number;
   [GL.TEXTURE_MAX_LOD]?: number;
   [GL.TEXTURE_COMPARE_FUNC]?: number; // COMPARE_FUNC);
-  // [GL.TEXTURE_MAX_ANISOTROPY]?: number; //
+  [GL.TEXTURE_MAX_ANISOTROPY_EXT]?: number; //
 };
 
 export type GLFunction =
